refactor(totals): consolidate table-calc totals into TotalQueryBuilder (PROD-8683)
Replace the four free functions in getTotalsQueryFromSource.ts (grand/column/row/
subtotal source→totals transforms) with a single TotalQueryBuilder class in the
QueryBuilder utils dir, mirroring MetricQueryBuilder's surface: args-object
constructor + one compileQuery() that dispatches on `kind`. Behaviour-identical
refactor — no functional change.
